requirements.txt
python-dotenv>=1.0.1
semantic-kernel
fastapi
requests
chainlit
numpy
pandas
openai
pymupdf
pillow
pydantic
json-repair
sqlalchemy
pyodbc
langchain
azure-identity
langchain-community 
azure-search-documents
opentelemetry-instrumentation-openai-v2
opentelemetry-instrumentation-langchain
